+++ fe/src/Games/Amazons.js
@@ -31,14 +31,6 @@ const initializeBoard = (config) => {
         [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
         [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
         [0, 0, 0, 1, 0, 0, 1, 0, 0, 0]],
-        "6_0": [
-            [0, 0, 1, 0, 0, 0],
-            [0, 0, 0, 0, 0, 0],
-            [0, 0, 0, 0, 0, 2],
-            [2, 0, 0, 0, 0, 0],
-            [0, 0, 0, 0, 0, 0],
-            [0, 0, 0, 1, 0, 0]
-        ],
     }
     if(!("size" in config && "variation" in config)){return false;}
     let key = config["size"].toString() + "_" + config["variation"].toString()