+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Re-organized the Do-list for progress
 - Added a header.
 - `<marquee> Added a footer. </marquee>`
+- Removed Amazons 6x6 configuration.
 # 2021-11-10
 - Found a name and domain for the project that I'm really happy with: strate.gg!
 - Wrote scripts for installing dependencies, preparing configurations and running servers